requirements.txt
aerich==0.7.1
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
dictdiffer==0.9.0
distro==1.8.0
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
magic-filter==1.0.9
multidict==6.0.4
netaddr==0.8.0
openai==0.27.6
psycopg==3.0.12
psycopg-binary==3.0.12
psycopg-pool==3.1.7
pydantic==1.10.7
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.29.0
tomlkit==0.11.8
tortoise-orm==0.19.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.9.2
yookassa==2.3.5
